Refactor ThingSpawner to be a singleton, plus add deletion of Things as they hit the left edge of the Board.

+++ SvS/TextWriter.cpp
@@ -16,7 +16,7 @@ void TextWriter::draw(sf::RenderWindow& w) {
 	w.draw(mText);
 }
 
-void TextWriter::setPosition(int x, int y) {
+void TextWriter::setPosition(float x, float y) {
 	mText.setPosition(x, y);
 }
